Add wercker CI support
Running `wercker dev` will set up a local dev env and watch files
for changes, upon which the tests are re-run. It's pretty nifty.
It also uses the same box for both `dev` and `build` pipelines,
so it should bring dev and build pretty close in terms of env
parity. That said – it doesn't matter much for a library of this
size, but it's nice.

No deploy targets yet; it may not get added it's since it's
simple enough to just `npm version <major|minor|patch>` and then
`npm publish` whenever a new version should be deployed.